Provide bytecode documentation.
